@@ -127,23 +127,24 @@
 		"node_modules"
 	],
 	"dependencies": {
-		"@angular/animations": "11.1.1",
-		"@angular/cdk": "11.1.1",
-		"@angular/common": "11.1.1",
-		"@angular/core": "11.1.1",
-		"@angular/flex-layout": "11.0.0-beta.33",
-		"@angular/forms": "11.1.1",
-		"@angular/material": "11.1.1",
-		"@angular/material-moment-adapter": "11.1.1",
-		"@angular/platform-browser": "11.1.1",
-		"@angular/platform-browser-dynamic": "11.1.1",
-		"@angular/router": "11.1.0",
-		"@cds/core": "5.0.0",
-		"@clr/angular": "5.0.0",
-		"@clr/core": "4.0.10",
-		"@clr/icons": "5.0.0",
-		"@clr/ui": "5.0.0",
-		"@fortawesome/angular-fontawesome": "0.8.1",
+		"@angular/animations": "11.2.4",
+		"@angular/cdk": "11.2.3",
+		"@angular/common": "11.2.4",
+		"@angular/core": "11.2.4",
+		"@angular/flex-layout": "^11.0.0-beta.33",
+		"@angular/forms": "^11.2.4",
+		"@angular/material": "^11.2.3",
+		"@angular/material-moment-adapter": "^11.2.3",
+		"@angular/platform-browser": "^11.2.4",
+		"@angular/platform-browser-dynamic": "^11.2.4",
+		"@angular/router": "^11.2.4",
+		"@cds/core": "^5.0.3",
+		"@clr/angular": "5.0.3",
+		"@clr/core": "4.0.12",
+		"@clr/icons": "5.0.3",
+		"@clr/ui": "5.0.3",
+		"@electron/remote": "^1.0.4",
+		"@fortawesome/angular-fontawesome": "0.8.2",
 		"@fortawesome/fontawesome-svg-core": "1.2.34",
 		"@fortawesome/free-brands-svg-icons": "5.15.2",
 		"@fortawesome/free-regular-svg-icons": "5.15.2",
@@ -183,58 +184,57 @@
 		"menu": "0.2.5",
 		"moment": "2.29.1",
 		"ngx-echarts": "6.0.0",
-		"ngx-electron": "2.2.0",
+		"ngx-electron": "^2.2.0",
 		"ngx-json-viewer": "3.0.2",
 		"ngx-lottie": "6.4.0",
-		"ngx-markdown": "11.0.1",
 		"ngx-mask": "11.1.4",
 		"ngx-order-pipe": "2.1.1",
 		"ngx-pagination": "5.0.0",
 		"ngx-toastr": "13.2.0",
-		"node-abi": "2.19.3",
+		"node-abi": "2.20.0",
 		"node-schedule": "2.0.0",
-		"open": "7.4.0",
+		"open": "8.0.2",
 		"popper.js": "1.16.1",
 		"portfinder": "1.0.28",
 		"primeicons": "4.1.0",
-		"primeng": "11.2.0",
+		"primeng": "11.3.0-rc.1",
 		"quill": "1.3.7",
 		"resize-observer-polyfill": "1.5.1",
-		"socket.io": "3.1.0",
-		"socket.io-client": "3.1.0",
+		"socket.io": "3.1.2",
+		"socket.io-client": "3.1.2",
 		"sudo-prompt": "9.2.1",
 		"tar": "6.1.0",
 		"text-mask-addons": "3.8.0",
 		"tslib": "2.1.0",
-		"zone.js": "0.11.3"
+		"zone.js": "0.11.4"
 	},
 	"devDependencies": {
-		"@angular-builders/custom-webpack": "11.0.0",
-		"@angular-devkit/build-angular": "0.1101.2",
-		"@angular/cli": "11.1.2",
-		"@angular/compiler": "11.1.1",
-		"@angular/compiler-cli": "11.1.1",
-		"@angular/language-service": "11.1.1",
+		"@angular-builders/custom-webpack": "^11.1.0",
+		"@angular-devkit/build-angular": "^0.1102.3",
+		"@angular/cli": "^11.2.3",
+		"@angular/compiler": "^11.2.4",
+		"@angular/compiler-cli": "^11.2.4",
+		"@angular/language-service": "^11.2.4",
 		"@electron/typescript-definitions": "8.8.0",
-		"@types/crypto-js": "4.0.1",
-		"@types/echarts": "4.9.3",
+		"@types/crypto-js": "^4.0.1",
+		"@types/echarts": "^4.9.4",
 		"@types/express": "4.17.11",
-		"@types/node": "14.14.22",
+		"@types/node": "14.14.31",
 		"@webcomponents/custom-elements": "1.4.3",
 		"codelyzer": "6.0.1",
-		"electron": "11.2.1",
-		"electron-builder": "22.9.1",
-		"electron-reload": "1.5.0",
+		"electron": "^12.0.0",
+		"electron-builder": "^22.10.5",
 		"electron-builder-notarize": "^1.2.0",
-		"eslint": "7.19.0",
+		"electron-reload": "^1.5.0",
+		"eslint": "^7.21.0",
 		"html-webpack-plugin": "4.5.1",
 		"npm-run-all": "4.1.5",
 		"protractor": "7.0.0",
-		"rxjs": "6.6.3",
-		"rxjs-compat": "6.6.3",
+		"rxjs": "^6.6.6",
+		"rxjs-compat": "^6.6.6",
 		"text-encoding-shim": "1.0.5",
 		"ts-node": "9.1.1",
-		"tslint": "6.1.3",
+		"tslint": "^6.1.3",
 		"typescript": "4.0.3",
 		"wait-on": "5.2.1",
 		"webpack": "4.46.0",
