@@ -5,15 +5,18 @@ const systemVersion = process.getSystemVersion();
 console.log(systemVersion);
 
 const {
-    app,
-    BrowserWindow,
-    Notification,
-    Menu,
-    protocol,
-    shell,
-    powerMonitor,
-    ipcMain
+  app,
+  BrowserWindow,
+  Notification,
+  Menu,
+  protocol,
+  shell,
+  powerMonitor,
+  ipcMain,
+  dialog,
 } = require('electron');
+
+require('@electron/remote/main').initialize();
 
 const path = require('path');
 const {version, productName, compilerVersion} = require('./package.json');
@@ -22,6 +25,7 @@ const {spawn} = require('child_process');
 const contextMenu = require('electron-context-menu');
 const AutoLaunch = require('auto-launch');
 const fs = require('fs');
+const keytar = require('keytar');
 
 const {TransitApiService} = require('./electron_modules/transit-api');
 const {SimpleosConnectService} = require('./electron_modules/simpleos-connect');
@@ -30,421 +34,445 @@ const {ClaimRewardsService} = require('./electron_modules/claim-rewards.js');
 
 class SimpleosWallet {
 
-    // wax auto claim manager (only loaded on DEFAULT compiler)
-    claimRW;
+  // wax auto claim manager (only loaded on DEFAULT compiler)
+  claimRW;
 
-    // ledger hardware wallet connector
-    ledger;
+  // ledger hardware wallet connector
+  ledger;
 
-    // transit api service
-    transit;
+  // transit api service
+  transit;
 
-    // simpleos connect service
-    connect;
+  // simpleos connect service
+  connect;
 
-    // chrome window object
-    win;
-    webContents;
+  // chrome window object
+  win;
+  webContents;
 
-    devtools;
-    serve;
-    debug;
-    isAutoLaunch;
-    deepLink;
-    isEnableAutoClaim = false;
-    PROTOCOL_PREFIX = 'simpleos';
-    simpleosAutoLauncher;
-    devMode = process.mainModule.filename.indexOf('app.asar') === -1;
-    loginOpts = app.getLoginItemSettings({
+  devtools;
+  serve;
+  debug;
+  isAutoLaunch;
+  deepLink;
+  isEnableAutoClaim = false;
+  PROTOCOL_PREFIX = 'simpleos';
+  simpleosAutoLauncher;
+  devMode = process.mainModule.filename.indexOf('app.asar') === -1;
+  loginOpts = app.getLoginItemSettings({
+    args: ['--autostart'],
+  });
+
+  constructor() {
+
+    if (compilerVersion === 'DEFAULT') {
+      this.simpleosAutoLauncher = new AutoLaunch({name: 'simpleos'});
+      this.claimRW = new ClaimRewardsService(this);
+      this.isEnableAutoClaim = this.claimRW.isEnableAutoClaim;
+    } else {
+      app.setLoginItemSettings({openAtLogin: false, args: ['--autostart']});
+      app.setLoginItemSettings({openAtLogin: false});
+    }
+  }
+
+  launchServices() {
+    // simpleos connect
+    this.connect = new SimpleosConnectService(this);
+    this.connect.init();
+    this.connect.startServer();
+
+    // transit api
+    this.transit = new TransitApiService(this);
+    this.transit.init();
+    this.transit.startServer();
+
+    // ledger integration
+    this.ledger = new LedgerManager(this);
+  }
+
+  init() {
+    app.allowRendererProcessReuse = true;
+    app.getVersion = () => version;
+    this.devtools = args.some(val => val === '--devtools');
+    this.debug = args.some(val => val === '--debug');
+    this.serve = args.some(val => val === '--serve');
+    this.isAutoLaunch = this.loginOpts.wasOpenedAtLogin || args.some(val => val === '--autostart');
+    if (this.claimRW) {
+      this.isEnableAutoClaim = this.claimRW.isEnableAutoClaim;
+      this.claimRW.writeLog(`Developer Mode: ${this.devMode}`);
+      console.log('Developer Mode:', this.devMode);
+
+      this.claimRW.writeLog(
+          `Auto Launcher: ${JSON.stringify(this.simpleosAutoLauncher)}`,
+      );
+      this.simpleosAutoLauncher.opts.appPath += ' --autostart';
+      this.simpleosAutoLauncher.isEnabled().then((status) => {
+        console.log('AUTO_LAUNCH:', status);
+        if (status) {
+          console.log('Auto launch already enabled!');
+          return;
+        }
+        if (!this.devMode) {
+          console.log('Enabling auto-launch!');
+          this.simpleosAutoLauncher.enable();
+        }
+      }).catch(function(err) {
+        console.log(err);
+      });
+      app.setLoginItemSettings({
+        openAtLogin: !this.devMode,
         args: ['--autostart'],
+      });
+    }
+
+    contextMenu();
+  }
+
+  runAutoLaunchMode() {
+    // check if another agent is running
+    console.log(`run Auto Launch Mode...`);
+    app.on('second-instance', (event, argv) => {
+      console.log(`check if another agent is running...`);
+
+      if (argv[1] === '--autostart') {
+        this.claimRW.writeLog(`Force quit agent in second instance...`);
+        app.quit();
+      }
     });
 
-    constructor() {
+    app.on('quit', () => {
+      this.claimRW.writeLog(`Quitting Agent...`);
+      this.claimRW.unlinkLALock();
+    });
 
-        if (compilerVersion === 'DEFAULT') {
-            this.simpleosAutoLauncher = new AutoLaunch({name: 'simpleos'});
-            this.claimRW = new ClaimRewardsService(this);
-            this.isEnableAutoClaim = this.claimRW.isEnableAutoClaim;
-        } else {
-            app.setLoginItemSettings({openAtLogin: false, args: ['--autostart']});
-            app.setLoginItemSettings({openAtLogin: false});
-        }
-    }
-
-    launchServices() {
-        // simpleos connect
-        this.connect = new SimpleosConnectService(this);
-        this.connect.init();
-        this.connect.startServer();
-
-        // transit api
-        this.transit = new TransitApiService(this);
-        this.transit.init();
-        this.transit.startServer();
-
-        // ledger integration
-        this.ledger = new LedgerManager(this);
-    }
-
-    init() {
-        app.allowRendererProcessReuse = true;
-        app.getVersion = () => version;
-        this.devtools = args.some(val => val === '--devtools');
-        this.debug = args.some(val => val === '--debug');
-        this.serve = args.some(val => val === '--serve');
-        this.isAutoLaunch = this.loginOpts.wasOpenedAtLogin || args.some(val => val === '--autostart');
-        if (this.claimRW) {
-            this.isEnableAutoClaim = this.claimRW.isEnableAutoClaim;
-            this.claimRW.writeLog(`Developer Mode: ${this.devMode}`);
-            console.log('Developer Mode:', this.devMode);
-
-            this.claimRW.writeLog(
-                `Auto Launcher: ${JSON.stringify(this.simpleosAutoLauncher)}`,
-            );
-            this.simpleosAutoLauncher.opts.appPath += ' --autostart';
-            this.simpleosAutoLauncher.isEnabled().then((status) => {
-                console.log('AUTO_LAUNCH:', status);
-                if (status) {
-                    console.log('Auto launch already enabled!');
-                    return;
-                }
-                if (!this.devMode) {
-                    console.log('Enabling auto-launch!');
-                    this.simpleosAutoLauncher.enable();
-                }
-            }).catch(function (err) {
-                console.log(err);
-            });
-            app.setLoginItemSettings({
-                openAtLogin: !this.devMode,
-                args: ['--autostart'],
-            });
-        }
-
-        contextMenu();
-    }
-
-    runAutoLaunchMode() {
-        // check if another agent is running
-        console.log(`run Auto Launch Mode...`);
-        app.on('second-instance', (event, argv) => {
-            console.log(`check if another agent is running...`);
-
-            if (argv[1] === '--autostart') {
-                this.claimRW.writeLog(`Force quit agent in second instance...`);
-                app.quit();
-            }
-        });
-
-        app.on('quit', () => {
-            this.claimRW.writeLog(`Quitting Agent...`);
-            this.claimRW.unlinkLALock();
-        });
-
-        app.on('ready', () => {
-            this.claimRW.unlinkLALock();
-            this.appendLock();
-            this.claimRW.autoClaimCheck();
-            console.log('READY!');
-            if (this.isEnableAutoClaim && productName === 'simpleos') {
-                this.claimRW.addTrayIcon();
-                this.claimRW.runAutoClaim();
-                if (process.platform === 'darwin') {
-                    app.dock.hide();
-                }
-            } else {
-                this.claimRW.writeLog(`Quitting disabled auto claim...`);
-                app.quit();
-            }
-            powerMonitor.on('suspend', () => {
-                this.claimRW.rescheduleAutoClaim();
-            });
-            powerMonitor.on('resume', () => {
-                this.claimRW.rescheduleAutoClaim();
-            });
-            powerMonitor.on('lock-screen', () => {
-                this.claimRW.rescheduleAutoClaim();
-            });
-        });
-    }
-
-    run() {
-        if (this.isAutoLaunch && this.claimRW) {
-            this.runAutoLaunchMode();
-        } else {
-            this.launchApp();
-            if (this.claimRW) {
-                this.claimRW.autoClaimCheck();
-                if (this.isEnableAutoClaim) {
-                    if (!(fs.existsSync(this.claimRW.lockAutoLaunchFile)) && productName ===
-                        'simpleos') {
-                        spawn(process.execPath, ['--autostart'], {
-                            detached: true,
-                            stdio: ['ignore', 'ignore', 'ignore'],
-                        }).unref();
-                    }
-                }
-            }
-        }
-    }
-
-    appendLock() {
-        if (this.isAutoLaunch) {
-            fs.writeFileSync(this.claimRW.lockAutoLaunchFile, process.pid);
-        } else {
-            fs.writeFileSync(this.claimRW.lockLaunchFile, process.pid);
-        }
-    }
-
-    getFocus() {
-        if (this.win) {
-            if (this.win.isMinimized()) {
-                this.win.restore();
-            }
-            this.win.focus();
-            this.win.show();
-
-        }
-
+    app.on('ready', () => {
+      this.claimRW.unlinkLALock();
+      this.appendLock();
+      this.claimRW.autoClaimCheck();
+      console.log('READY!');
+      if (this.isEnableAutoClaim && productName === 'simpleos') {
+        this.claimRW.addTrayIcon();
+        this.claimRW.runAutoClaim();
         if (process.platform === 'darwin') {
-            app.dock.hide();
-            this.win.setAlwaysOnTop(false);
-            this.win.setVisibleOnAllWorkspaces(true);
-            this.win.setFullScreenable(false);
-            app.dock.show().catch(console.log);
+          app.dock.hide();
         }
-    }
+      } else {
+        this.claimRW.writeLog(`Quitting disabled auto claim...`);
+        app.quit();
+      }
+      powerMonitor.on('suspend', () => {
+        this.claimRW.rescheduleAutoClaim();
+      });
+      powerMonitor.on('resume', () => {
+        this.claimRW.rescheduleAutoClaim();
+      });
+      powerMonitor.on('lock-screen', () => {
+        this.claimRW.rescheduleAutoClaim();
+      });
+    });
+  }
 
-    unfocus() {
-        switch (process.platform) {
-            case 'win32': {
-                this.win.minimize();
-                break;
-            }
-            case 'linux': {
-                this.win.hide();
-                break;
-            }
-            case 'darwin': {
-                if (this.claimRW) {
-                    this.claimRW.writeLog(`unfocus `);
-                }
-                Menu.sendActionToFirstResponder('hide:');
-                break;
-            }
+  run() {
+    if (this.isAutoLaunch && this.claimRW) {
+      this.runAutoLaunchMode();
+    } else {
+      this.launchApp();
+      if (this.claimRW) {
+        this.claimRW.autoClaimCheck();
+        if (this.isEnableAutoClaim) {
+          if (!(fs.existsSync(this.claimRW.lockAutoLaunchFile)) && productName ===
+              'simpleos') {
+            spawn(process.execPath, ['--autostart'], {
+              detached: true,
+              stdio: ['ignore', 'ignore', 'ignore'],
+            }).unref();
+          }
         }
+      }
+    }
+  }
+
+  appendLock() {
+    const pidString = process.pid.toString(10);
+    if (this.isAutoLaunch) {
+      fs.writeFileSync(this.claimRW.lockAutoLaunchFile, pidString);
+    } else {
+      fs.writeFileSync(this.claimRW.lockLaunchFile, pidString);
+    }
+  }
+
+  getFocus() {
+    if (this.win) {
+      if (this.win.isMinimized()) {
+        this.win.restore();
+      }
+      this.win.focus();
+      this.win.show();
+
     }
 
-    regURI() {
-        app.setAsDefaultProtocolClient(this.PROTOCOL_PREFIX);
-        protocol.registerHttpProtocol(this.PROTOCOL_PREFIX, (req, callback) => {
-            if (req.url < 128) {
-                this.deepLink = req;
-                setTimeout(() => {
-                    this.win.webContents.send('request', {
-                        message: 'launch',
-                        content: this.deepLink.url,
-                    });
-                    this.win.webContents.send('electron', {message: 'type', content: process.platform});
-                }, 5000);
-            }
-            callback();
-        });
+    if (process.platform === 'darwin') {
+      app.dock.hide();
+      this.win.setAlwaysOnTop(false);
+      this.win.setVisibleOnAllWorkspaces(true);
+      this.win.setFullScreenable(false);
+      app.dock.show().catch(console.log);
     }
+  }
 
-    launchApp() {
+  unfocus() {
+    switch (process.platform) {
+      case 'win32': {
+        this.win.minimize();
+        break;
+      }
+      case 'linux': {
+        this.win.hide();
+        break;
+      }
+      case 'darwin': {
         if (this.claimRW) {
-            const gotTheLock = app.requestSingleInstanceLock();
-            this.claimRW.writeLog(`On Launching File LAUNCH: ${(fs.existsSync(
-                this.claimRW.lockLaunchFile))} | The LOCK: ${gotTheLock}`);
-            if (fs.existsSync(this.claimRW.lockLaunchFile)) {
-                if (gotTheLock) {
-                    this.claimRW.unlinkLLock();
-                } else {
-                    console.log('quiting');
-                    app.quit();
-                    return;
-                }
-            }
-
-            this.appendLock();
-
-            app.on('second-instance', () => {
-                console.log('launching second instance...');
-                if (this.win) {
-                    if (this.win.isMinimized()) {
-                        this.win.restore();
-                    }
-                    this.win.focus();
-                }
-            });
+          this.claimRW.writeLog(`unfocus `);
         }
-
-        app.on('ready', () => {
-            console.log('Electron ready');
-            this.launchServices();
-            this.createWindow().catch(console.log);
-        });
-
-        app.on('window-all-closed', () => {
-            if (this.claimRW) {
-                this.claimRW.writeLog(`Quitting Application...`);
-                this.claimRW.clearLock();
-                this.claimRW.unlinkLLock();
-            }
-            app.quit();
-        });
-
-        app.on('activate', async () => {
-            if (this.win === null) {
-                await this.createWindow();
-            }
-        });
-
-        app.on('will-finish-launching', () => {
-            app.on('open-url', (e, url) => {
-                //e.preventDefault();
-                console.log(url);
-            });
-        });
+        Menu.sendActionToFirstResponder('hide:');
+        break;
+      }
     }
+  }
 
-    notifyTrx(title, body, autoClose, trx_id) {
-        const notification = new Notification({
-            title: title,
-            body: body,
-        });
-        notification.show();
-        notification.on('click', () => {
-            shell.openExternal('https://wax.bloks.io/transaction/' + trx_id).catch(console.log);
-        });
-        if (autoClose > 0) {
-            setTimeout(() => {
-                notification.close();
-            }, autoClose);
-        }
-    }
+  regURI() {
+    app.setAsDefaultProtocolClient(this.PROTOCOL_PREFIX);
+    protocol.registerHttpProtocol(this.PROTOCOL_PREFIX, (req, callback) => {
+      if (req.url < 128) {
+        this.deepLink = req;
+        setTimeout(() => {
+          this.win.webContents.send('request', {
+            message: 'launch',
+            content: this.deepLink.url,
+          });
+          this.win.webContents.send('electron', {message: 'type', content: process.platform});
+        }, 5000);
+      }
+      callback();
+    });
+  }
 
-   async notify(title, body, autoClose) {
-        const notification = new Notification({
-            title: title,
-            body: body,
-        });
-        notification.show();
-        notification.on('click', () => {
-            if (this.win) {
-                this.getFocus();
-            }
-        });
-        if (autoClose > 0) {
-            setTimeout(() => {
-                notification.close();
-            }, autoClose);
-        }
-    }
-
-    async createWindow() {
-
-        this.regURI();
-
-        let _icon = path.join(__dirname, 'build/icons/simpleos/icon.png');
-        let _bgColor = '#222222';
-        if (compilerVersion === 'LIBERLAND') {
-            _icon = path.join(__dirname, 'build/icons/liberland/icon.png');
-            _bgColor = '#2a566f';
-        }
-
-        if (!fs.existsSync(_icon)) {
-            console.log('failed to load icon file');
-        }
-
-        this.win = new BrowserWindow({
-            title: productName,
-            show: false,
-            paintWhenInitiallyHidden: true,
-            titleBarStyle: 'hiddenInset',
-            webPreferences: {
-                nodeIntegration: true,
-                nodeIntegrationInWorker: true,
-                webSecurity: !this.serve,
-                devTools: this.devtools,
-                enableRemoteModule: true,
-                contextIsolation: false
-            },
-            darkTheme: true,
-            width: 1440,
-            height: 920,
-            minWidth: 1024,
-            minHeight: 600,
-            backgroundColor: _bgColor,
-            frame: false,
-            enableRemoteModule: false,
-        });
-
-        this.webContents = this.win.webContents;
-
-        this.webContents.on('did-finish-load', (e) => {
-            console.log('did-finish-load');
-            this.win.setTitle(productName);
-            this.win.setIcon(_icon);
-        });
-
-        ipcMain.on('electron', (event, args) => {
-            if (args === 'request_os') {
-                this.win.webContents.send('electron', {
-                    event: 'platform_reply',
-                    content: process.platform
-                });
-            }
-        });
-
-        this.win.once('ready-to-show', () => {
-            console.log('window is ready to show');
-            this.win.show();
-        });
-
-        // win.removeMenu();
-
-        console.log('SERVE:', this.serve);
-        if (this.serve) {
-            require('electron-reload')(__dirname, {
-                electron: path.join(__dirname, 'node_modules', '.bin', 'electron'),
-                hardResetMethod: 'exit',
-            });
-            await this.win.loadURL('http://localhost:7777');
+  launchApp() {
+    if (this.claimRW) {
+      const gotTheLock = app.requestSingleInstanceLock();
+      this.claimRW.writeLog(`On Launching File LAUNCH: ${(fs.existsSync(
+          this.claimRW.lockLaunchFile))} | The LOCK: ${gotTheLock}`);
+      if (fs.existsSync(this.claimRW.lockLaunchFile)) {
+        if (gotTheLock) {
+          this.claimRW.unlinkLLock();
         } else {
-            await this.win.loadURL(url.format({
-                pathname: path.join(__dirname, 'ng-dist/index.html'),
-                protocol: 'file:',
-                slashes: true,
-            }));
+          console.log('quiting');
+          app.quit();
+          return;
         }
+      }
 
-        // catch console logs from angular app
-        if (this.debug) {
-            this.webContents.on('console-message', (e, level, msg, line) => {
-                if (level === 1) {
-                    console.log(`Log [${line}]: ${msg}`);
-                }
-            });
+      this.appendLock();
+
+      app.on('second-instance', () => {
+        console.log('launching second instance...');
+        if (this.win) {
+          if (this.win.isMinimized()) {
+            this.win.restore();
+          }
+          this.win.focus();
         }
-
-
-        console.log('DEVTOOLS:', this.devtools);
-        if (this.devtools) {
-            this.webContents.openDevTools();
-        } else {
-            this.webContents.on('devtools-opened', () => {
-                console.log('devtools opened');
-                this.webContents.closeDevTools();
-            });
-        }
-
-        this.win.on('closed', () => {
-            this.win = null;
-        });
-
+      });
     }
+
+    app.on('ready', () => {
+      console.log('Electron ready');
+      this.launchServices();
+      this.createWindow().catch(console.log);
+    });
+
+    app.on('window-all-closed', () => {
+      if (this.claimRW) {
+        this.claimRW.writeLog(`Quitting Application...`);
+        this.claimRW.clearLock();
+        this.claimRW.unlinkLLock();
+      }
+      app.quit();
+    });
+
+    app.on('activate', async () => {
+      if (this.win === null) {
+        await this.createWindow();
+      }
+    });
+
+    app.on('will-finish-launching', () => {
+      app.on('open-url', (e, url) => {
+        //e.preventDefault();
+        console.log(url);
+      });
+    });
+  }
+
+  notifyTrx(title, body, autoClose, trx_id) {
+    const notification = new Notification({
+      title: title,
+      body: body,
+    });
+    notification.show();
+    notification.on('click', () => {
+      shell.openExternal('https://wax.bloks.io/transaction/' + trx_id).catch(console.log);
+    });
+    if (autoClose > 0) {
+      setTimeout(() => {
+        notification.close();
+      }, autoClose);
+    }
+  }
+
+  async notify(title, body, autoClose) {
+    const notification = new Notification({
+      title: title,
+      body: body,
+    });
+    notification.show();
+    notification.on('click', () => {
+      if (this.win) {
+        this.getFocus();
+      }
+    });
+    if (autoClose > 0) {
+      setTimeout(() => {
+        notification.close();
+      }, autoClose);
+    }
+  }
+
+  async createWindow() {
+
+    this.regURI();
+
+    let _icon = path.join(__dirname, 'build/icons/simpleos/icon.png');
+    let _bgColor = '#222222';
+    if (compilerVersion === 'LIBERLAND') {
+      _icon = path.join(__dirname, 'build/icons/liberland/icon.png');
+      _bgColor = '#2a566f';
+    }
+
+    if (!fs.existsSync(_icon)) {
+      console.log('failed to load icon file');
+    }
+
+    this.win = new BrowserWindow({
+      title: productName,
+      show: false,
+      paintWhenInitiallyHidden: true,
+      titleBarStyle: 'hiddenInset',
+      webPreferences: {
+        nodeIntegration: true,
+        nodeIntegrationInWorker: true,
+        webSecurity: !this.serve,
+        devTools: this.devtools,
+        enableRemoteModule: true,
+        contextIsolation: false,
+      },
+      darkTheme: true,
+      width: 1440,
+      height: 920,
+      minWidth: 1024,
+      minHeight: 600,
+      backgroundColor: _bgColor,
+      frame: false,
+    });
+
+    this.webContents = this.win.webContents;
+
+    this.webContents.on('did-finish-load', (e) => {
+      console.log('did-finish-load');
+      this.win.setTitle(productName);
+      this.win.setIcon(_icon);
+    });
+
+    ipcMain.on('electron', (event, args) => {
+      if (args === 'request_os') {
+        this.win.webContents.send('electron', {
+          event: 'platform_reply',
+          content: process.platform,
+        });
+      }
+    });
+
+    // native prompt for export folder
+    ipcMain.handle('read-export-dir', async (event, filename) => {
+      const dirs = await dialog.showOpenDialog({properties: ['openDirectory']});
+      if (dirs.filePaths.length > 0) {
+        return path.join(dirs.filePaths[0], filename);
+      } else {
+        return null;
+      }
+    });
+
+    // native prompt for open file
+    ipcMain.handle('read-open-file', async (event, filename) => {
+      const selected = await dialog.showOpenDialog({properties: ['openFile']});
+      if (selected.filePaths.length > 0) {
+        return selected.filePaths[0];
+      } else {
+        return null;
+      }
+    });
+
+    // extract key from local keytar
+    ipcMain.handle('keytar-getPassword', async (event, key) => {
+      return await keytar.getPassword('simpleos', key);
+    });
+
+    this.win.once('ready-to-show', () => {
+      console.log('window is ready to show');
+      this.win.show();
+    });
+
+    // win.removeMenu();
+
+    console.log('SERVE:', this.serve);
+    if (this.serve) {
+      require('electron-reload')(__dirname, {
+        electron: path.join(__dirname, 'node_modules', '.bin', 'electron'),
+        hardResetMethod: 'exit',
+      });
+      await this.win.loadURL('http://localhost:7777');
+    } else {
+      await this.win.loadURL(url.format({
+        pathname: path.join(__dirname, 'ng-dist/index.html'),
+        protocol: 'file:',
+        slashes: true,
+      }));
+    }
+
+    // catch console logs from angular app
+    if (this.debug) {
+      this.webContents.on('console-message', (e, level, msg, line) => {
+        if (level === 1) {
+          console.log(`Log [${line}]: ${msg}`);
+        }
+      });
+    }
+
+    console.log('DEVTOOLS:', this.devtools);
+    if (this.devtools) {
+      this.webContents.openDevTools();
+    } else {
+      this.webContents.on('devtools-opened', () => {
+        console.log('devtools opened');
+        this.webContents.closeDevTools();
+      });
+    }
+
+    this.win.on('closed', () => {
+      this.win = null;
+    });
+
+  }
 }
 
 const wallet = new SimpleosWallet();
