@@ -17,634 +17,637 @@ import {ThemeService} from './services/theme.service';
 import {Title} from '@angular/platform-browser';
 import {LedgerService} from './services/ledger/ledger.service';
 import {NotificationService} from './services/notification.service';
-import {EOSAccount} from "./interfaces/account";
-import {HttpClient, HttpHeaders} from "@angular/common/http";
+import {EOSAccount} from './interfaces/account';
+import {HttpClient, HttpHeaders} from '@angular/common/http';
+
+import {BrowserWindow} from '@electron/remote';
+
 
 export interface LedgerSlot {
-    publicKey: string;
-    account: string;
+	publicKey: string;
+	account: string;
 }
 
 // @ts-ignore
 @Component({
-    selector: 'app-root',
-    templateUrl: './app.component.html',
-    styleUrls: ['./app.component.css'],
+	selector: 'app-root',
+	templateUrl: './app.component.html',
+	styleUrls: ['./app.component.css'],
 })
 export class AppComponent implements OnInit, AfterViewInit {
 
-    confirmForm: FormGroup;
-    wrongpass = '';
-    txerror = '';
-    transitconnect = false;
-    externalActionModal = false;
-    dapp_name = '';
-    selectedAccount = new BehaviorSubject<any>('');
-    accountChange: Subscription;
-    ledgerOpen: boolean;
-    update: boolean;
-    busy: boolean;
-    action_json: any[];
-    accSlots: LedgerSlot[];
-    selectedSlot: LedgerSlot;
-    selectedSlotIndex: number;
-    dnSet = false;
-    activeChain = null;
+	constructor(
+		private fb: FormBuilder,
+		public network: NetworkService,
+		public ledger: LedgerService,
+		public aService: AccountsService,
+		private titleService: Title,
+		private eosjs: Eosjs2Service,
+		private crypto: CryptoService,
+		private connect: ConnectService,
+		private router: Router,
+		private autobackup: BackupService,
+		private trxFactory: TransactionFactoryService,
+		private zone: NgZone,
+		private cdr: ChangeDetectorRef,
+		private _electronService: ElectronService,
+		public theme: ThemeService,
+		private toaster: NotificationService,
+		private http: HttpClient
+	) {
 
-    public version = environment.VERSION;
-    public compilerVersion = environment.COMPILERVERSION;
+		console.log('BW', BrowserWindow);
 
-    updateauthWarning = false;
-    public newVersion: any;
-    public external_signer: string;
-    private fullTrxData: any;
-    private replyEvent: any;
+		if (this.compilerVersion === 'LIBERLAND') {
+			this.titleService.setTitle('Liberland Wallet v' + this.version);
+			this.theme.liberlandTheme();
 
-    public isMac: boolean;
+			if (!this.network.activeChain.name.startsWith('LIBERLAND')) {
+				this.activeChain = this.network.defaultChains.find((chain) => chain.name === 'LIBERLAND TESTNET');
+				localStorage.setItem('simplEOS.activeChainID', this.activeChain.id);
+				this.network.changeChain(this.activeChain.id).catch(console.log);
+			}
 
-    public transitEventHandler: any;
-    private eventFired: boolean;
-    public loadingTRX: boolean;
-    private transitMode = false;
-    private isSimpleosConnect: boolean;
+		} else {
+			this.theme.defaultTheme();
+			this.titleService.setTitle('SimplEOS Wallet v' + this.version);
+		}
 
-    constructor(
-        private fb: FormBuilder,
-        public network: NetworkService,
-        public ledger: LedgerService,
-        public aService: AccountsService,
-        private titleService: Title,
-        private eosjs: Eosjs2Service,
-        private crypto: CryptoService,
-        private connect: ConnectService,
-        private router: Router,
-        private autobackup: BackupService,
-        private trxFactory: TransactionFactoryService,
-        private zone: NgZone,
-        private cdr: ChangeDetectorRef,
-        private _electronService: ElectronService,
-        public theme: ThemeService,
-        private toaster: NotificationService,
-        private http: HttpClient
-    ) {
+		this.confirmForm = this.fb.group({
+			pass: ['', Validators.required],
+		});
 
-        if (this.compilerVersion === 'LIBERLAND') {
-            this.titleService.setTitle('Liberland Wallet v' + this.version);
-            this.theme.liberlandTheme();
+		// wait 30 seconds to automatic backup
+		this.autobackup.startTimeout();
 
-            if (!this.network.activeChain.name.startsWith('LIBERLAND')) {
-                this.activeChain = this.network.defaultChains.find((chain) => chain.name === 'LIBERLAND TESTNET');
-                localStorage.setItem('simplEOS.activeChainID', this.activeChain.id);
-                this.network.changeChain(this.activeChain.id).catch(console.log);
-            }
+		this.accSlots = [];
+		this.selectedSlot = null;
+		this.selectedSlotIndex = null;
+		this.update = false;
+		this.newVersion = null;
+		this.aService.versionSys = this.version;
 
-        } else {
-            this.theme.defaultTheme();
-            this.titleService.setTitle('SimplEOS Wallet v' + this.version);
-        }
+		this.ledgerOpen = false;
+		this.loadingTRX = false;
 
-        this.confirmForm = this.fb.group({
-            pass: ['', Validators.required],
-        });
+		this.busy = false;
+		this.ledger.startListener();
+		this.startIPCListeners();
+	}
 
-        // wait 30 seconds to automatic backup
-        this.autobackup.startTimeout();
+	get maximized(): boolean {
+		return this._electronService.remote.getCurrentWindow().isMaximized();
+	}
 
-        this.accSlots = [];
-        this.selectedSlot = null;
-        this.selectedSlotIndex = null;
-        this.update = false;
-        this.newVersion = null;
-        this.aService.versionSys = this.version;
+	get requiredLedgerInfo() {
+		return {
+			device: this.crypto.requiredLedgerDevice,
+			slot: this.crypto.requiredLedgerSlot
+		};
+	}
 
-        this.ledgerOpen = false;
-        this.loadingTRX = false;
+	confirmForm: FormGroup;
+	wrongpass = '';
+	txerror = '';
+	transitconnect = false;
+	externalActionModal = false;
+	dapp_name = '';
+	selectedAccount = new BehaviorSubject<any>('');
+	accountChange: Subscription;
+	ledgerOpen: boolean;
+	update: boolean;
+	busy: boolean;
+	action_json: any[];
+	accSlots: LedgerSlot[];
+	selectedSlot: LedgerSlot;
+	selectedSlotIndex: number;
+	dnSet = false;
+	activeChain = null;
 
-        this.busy = false;
-        this.ledger.startListener();
-        this.startIPCListeners();
-    }
+	public version = environment.VERSION;
+	public compilerVersion = environment.COMPILERVERSION;
 
-    ngOnInit(): void {
-        this.connect.ipc.send('electron', 'request_os');
-    }
+	updateauthWarning = false;
+	public newVersion: any;
+	public external_signer: string;
+	private fullTrxData: any;
+	private replyEvent: any;
 
-    ngAfterViewInit() {
+	public isMac: boolean;
 
-        const dnSavedValue = localStorage.getItem('use_light_theme');
-        if (dnSavedValue && dnSavedValue === 'true') {
-            this.dnSet = true;
-            this.theme.lightMode = true;
-            this.theme.lightTheme();
-            this.cdr.detectChanges();
-        }
+	public transitEventHandler: any;
+	private eventFired: boolean;
+	public loadingTRX: boolean;
+	private transitMode = false;
+	private isSimpleosConnect: boolean;
 
-        setTimeout(() => {
+	mode = 'local';
 
-            this.network.connect(false).catch(console.log);
+	ngOnInit(): void {
+		this.connect.ipc.send('electron', 'request_os');
+	}
 
-            setTimeout(async () => {
-                if (this.compilerVersion === 'DEFAULT') {
-                    this.newVersion = await this.eosjs.checkSimpleosUpdate();
-                } else {
-                    try{
-                        const results = await this.http.get('https://raw.githubusercontent.com/eosrio/simpleos/master/latest.json', {
-                            headers: new HttpHeaders({
-                                'Cache-Control': 'no-cache, no-store, must-revalidate, post-check=0, pre-check=0',
-                                'Pragma': 'no-cache',
-                                'Expires': '0'
-                            })
-                        }).toPromise();
-                        if (results && results[this.compilerVersion]) {
-                            this.newVersion = results[this.compilerVersion];
-                        }
-                    }catch (e) {
-                        console.log('Failed to check version in github');
-                    }
+	ngAfterViewInit() {
 
-                }
-                if (this.newVersion) {
-                    const remoteVersionNum = parseInt(this.newVersion['version_number'].replace(/[v.]/g, ''));
-                    const currentVersionNum = parseInt(this.version.replace(/[.]/g, ''));
-                    console.log(`Remote Version: ${remoteVersionNum}`);
-                    console.log(`Local Version: ${currentVersionNum}`);
-                    if (remoteVersionNum > currentVersionNum) {
-                        this.update = true;
-                    }
-                }
-            }, 5000);
+		const dnSavedValue = localStorage.getItem('use_light_theme');
+		if (dnSavedValue && dnSavedValue === 'true') {
+			this.dnSet = true;
+			this.theme.lightMode = true;
+			this.theme.lightTheme();
+			this.cdr.detectChanges();
+		}
 
-        }, 500);
-    }
+		setTimeout(() => {
 
-    startIPCListeners() {
-        if (this.connect.ipc) {
+			this.network.connect(false).catch(console.log);
 
-            // Bind simpleos connect requests
-            this.onSimpleosConnectMessage = this.onSimpleosConnectMessage.bind(this);
-            this.connect.ipc.on('sc_request', this.onSimpleosConnectMessage);
+			setTimeout(async () => {
+				if (this.compilerVersion === 'DEFAULT') {
+					this.newVersion = await this.eosjs.checkSimpleosUpdate();
+				} else {
+					try {
+						const results = await this.http.get('https://raw.githubusercontent.com/eosrio/simpleos/master/latest.json', {
+							headers: new HttpHeaders({
+								'Cache-Control': 'no-cache, no-store, must-revalidate, post-check=0, pre-check=0',
+								'Pragma': 'no-cache',
+								'Expires': '0'
+							})
+						}).toPromise();
+						if (results && results[this.compilerVersion]) {
+							this.newVersion = results[this.compilerVersion];
+						}
+					} catch (e) {
+						console.log('Failed to check version in github');
+					}
 
-            // Bind transit api requests
-            this.onTransitApiMessage = this.onTransitApiMessage.bind(this);
-            this.connect.ipc.on('request', this.onTransitApiMessage);
+				}
+				if (this.newVersion) {
+					const remoteVersionNum = parseInt(this.newVersion['version_number'].replace(/[v.]/g, ''), 10);
+					const currentVersionNum = parseInt(this.version.replace(/[.]/g, ''), 10);
+					console.log(`Remote Version: ${remoteVersionNum}`);
+					console.log(`Local Version: ${currentVersionNum}`);
+					if (remoteVersionNum > currentVersionNum) {
+						this.update = true;
+					}
+				}
+			}, 5000);
 
-            // Bind transit api requests
-            this.onElectron = this.onElectron.bind(this);
-            this.connect.ipc.on('electron', this.onElectron);
-        }
-    }
+		}, 500);
+	}
 
-    private onElectron(event, payload) {
-        if (event) {
-            if (payload.event === 'platform_reply') {
-                console.log('Platform:', payload.content);
-                this.isMac = payload.content === 'darwin';
-                this.cdr.detectChanges();
-            }
-        }
-    }
+	startIPCListeners() {
+		if (this.connect.ipc) {
 
-    private onTransitApiMessage(event, payload) {
-        this.transitEventHandler = event;
-        switch (payload.message) {
-            case 'launch': {
-                console.log(payload);
-                break;
-            }
-            case 'accounts': {
-                event.sender.send('accountsResponse', this.aService.accounts.map(a => a.name));
-                break;
-            }
-            case 'connect': {
-                this.dapp_name = payload.content.appName;
-                const result = this.changeChain(payload.content.chainId);
-                event.sender.send('connectResponse', result);
-                break;
-            }
-            case 'login': {
-                if (localStorage.getItem('simpleos-hash') && this.crypto.getLockStatus()) {
-                    this.eventFired = true;
-                    this.transitEventHandler.sender.send('loginResponse', {status: 'CANCELLED'});
-                    return;
-                }
-                const reqAccount = payload.content.account;
-                if (reqAccount) {
-                    const foundAccount = this.aService.accounts.find((a: EOSAccount) => a.name === reqAccount);
-                    if (foundAccount) {
-                        this.selectedAccount.next(foundAccount);
-                        event.sender.send('loginResponse', foundAccount);
-                        break;
-                    } else {
-                        console.log('Account not imported on wallet!');
-                        event.sender.send('loginResponse', {});
-                    }
-                    return;
-                }
-                this.zone.run(() => {
-                    this.isSimpleosConnect = false;
-                    this.transitconnect = true;
-                    this.eventFired = false;
-                });
-                if (!this.aService.accounts) {
-                    console.log('No account found!');
-                    event.sender.send('loginResponse', {});
-                }
-                if (this.aService.accounts.length > 0) {
-                    this.accountChange = this.selectedAccount.subscribe((data) => {
-                        if (data) {
-                            event.sender.send('loginResponse', data);
-                            this.accountChange.unsubscribe();
-                            this.selectedAccount.next(null);
-                        }
-                    });
-                } else {
-                    console.log('No account found!');
-                    event.sender.send('loginResponse', {});
-                }
-                break;
-            }
-            case 'logout': {
-                const reqAccount = payload.content.account;
-                console.log(reqAccount);
-                this.dapp_name = null;
-                event.sender.send('logoutResponse', {});
-                break;
-            }
-            case 'disconnect': {
-                this.dapp_name = null;
-                event.sender.send('disconnectResponse', {});
-                break;
-            }
-            case 'publicKeys': {
-                const localKeys = JSON.parse(localStorage.getItem('eos_keys.' + this.eosjs.chainId));
-                event.sender.send('publicKeyResponse', Object.keys(localKeys));
-                break;
-            }
-            case 'sign': {
-                this.transitMode = true;
-                this.onTransitSign(event, payload).catch(console.log);
-                break;
-            }
-            default: {
-                console.log(payload);
-            }
-        }
-    }
+			// Bind simpleos connect requests
+			this.onSimpleosConnectMessage = this.onSimpleosConnectMessage.bind(this);
+			this.connect.ipc.on('sc_request', this.onSimpleosConnectMessage);
 
-    private onSimpleosConnectMessage(event, payload) {
-        switch (payload.message) {
-            case 'change_chain': {
-                if (payload.chain_id !== this.network.activeChain.id) {
-                    const result: boolean = this.changeChain(payload.chain_id);
-                    // wait for chain to change
-                    const onceListener = this.network.networkingReady.subscribe((value) => {
-                        if (value) {
-                            event.sender.send('changeChainResponse', result);
-                            onceListener.unsubscribe();
-                        }
-                    });
-                } else {
-                    event.sender.send('changeChainResponse', true);
-                }
-                break;
-            }
-            case 'authorizations': {
-                this.toaster.onInfo( 'wallet connection','external application requested public account info');
-                this.transitconnect = true;
-                this.isSimpleosConnect = true;
+			// Bind transit api requests
+			this.onTransitApiMessage = this.onTransitApiMessage.bind(this);
+			this.connect.ipc.on('request', this.onTransitApiMessage);
 
-                if (this.aService.accounts.length > 0) {
-                    this.accountChange = this.selectedAccount.subscribe((data) => {
-                        if (data) {
-                            event.sender.send('authorizationsResponse', data);
-                            this.accountChange.unsubscribe();
-                            this.selectedAccount.next(null);
-                            this.transitconnect = false;
-                        }
-                    });
-                } else {
-                    console.log('No account found!');
-                    event.sender.send('authorizationsResponse', {});
-                }
+			// Bind transit api requests
+			this.onElectron = this.onElectron.bind(this);
+			this.connect.ipc.on('electron', this.onElectron);
+		}
+	}
 
-                break;
-            }
-            case 'sign': {
-                if (this.crypto.getLockStatus()) {
-                    event.sender.send('signResponse', {
-                        status: 'CANCELLED'
-                    });
-                } else {
-                    this.simpleosConnectSign(event, payload).catch(console.log);
-                }
-                break;
-            }
-        }
-    }
+	private onElectron(event, payload) {
+		if (event) {
+			if (payload.event === 'platform_reply') {
+				console.log('Platform:', payload.content);
+				this.isMac = payload.content === 'darwin';
+				this.cdr.detectChanges();
+			}
+		}
+	}
 
-    private changeChain(chainId): boolean {
-        if (this.network.defaultChains.find((chain) => chain.id === chainId)) {
-            if (this.network.activeChain.id !== chainId) {
-                this.network.changeChain(chainId).catch(console.log);
-            }
-            return true;
-        } else {
-            return false;
-        }
-    }
+	private onTransitApiMessage(event, payload) {
+		this.transitEventHandler = event;
+		switch (payload.message) {
+			case 'launch': {
+				console.log(payload);
+				break;
+			}
+			case 'accounts': {
+				event.sender.send('accountsResponse', this.aService.accounts.map(a => a.name));
+				break;
+			}
+			case 'connect': {
+				this.dapp_name = payload.content.appName;
+				const result = this.changeChain(payload.content.chainId);
+				event.sender.send('connectResponse', result);
+				break;
+			}
+			case 'login': {
+				if (localStorage.getItem('simpleos-hash') && this.crypto.getLockStatus()) {
+					this.eventFired = true;
+					this.transitEventHandler.sender.send('loginResponse', {status: 'CANCELLED'});
+					return;
+				}
+				const reqAccount = payload.content.account;
+				if (reqAccount) {
+					const foundAccount = this.aService.accounts.find((a: EOSAccount) => a.name === reqAccount);
+					if (foundAccount) {
+						this.selectedAccount.next(foundAccount);
+						event.sender.send('loginResponse', foundAccount);
+						break;
+					} else {
+						console.log('Account not imported on wallet!');
+						event.sender.send('loginResponse', {});
+					}
+					return;
+				}
+				this.zone.run(() => {
+					this.isSimpleosConnect = false;
+					this.transitconnect = true;
+					this.eventFired = false;
+				});
+				if (!this.aService.accounts) {
+					console.log('No account found!');
+					event.sender.send('loginResponse', {});
+				}
+				if (this.aService.accounts.length > 0) {
+					this.accountChange = this.selectedAccount.subscribe((data) => {
+						if (data) {
+							event.sender.send('loginResponse', data);
+							this.accountChange.unsubscribe();
+							this.selectedAccount.next(null);
+						}
+					});
+				} else {
+					console.log('No account found!');
+					event.sender.send('loginResponse', {});
+				}
+				break;
+			}
+			case 'logout': {
+				const reqAccount = payload.content.account;
+				console.log(reqAccount);
+				this.dapp_name = null;
+				event.sender.send('logoutResponse', {});
+				break;
+			}
+			case 'disconnect': {
+				this.dapp_name = null;
+				event.sender.send('disconnectResponse', {});
+				break;
+			}
+			case 'publicKeys': {
+				const localKeys = JSON.parse(localStorage.getItem('eos_keys.' + this.eosjs.chainId));
+				event.sender.send('publicKeyResponse', Object.keys(localKeys));
+				break;
+			}
+			case 'sign': {
+				this.transitMode = true;
+				this.onTransitSign(event, payload).catch(console.log);
+				break;
+			}
+			default: {
+				console.log(payload);
+			}
+		}
+	}
 
-    onLoginModalClose(ev) {
-        if (this.transitEventHandler && ev === false && this.eventFired === false) {
-            this.eventFired = true;
-            this.transitEventHandler.sender.send('loginResponse', {status: 'CANCELLED'});
-        }
-    }
+	private onSimpleosConnectMessage(event, payload) {
+		switch (payload.message) {
+			case 'change_chain': {
+				if (payload.chain_id !== this.network.activeChain.id) {
+					const result: boolean = this.changeChain(payload.chain_id);
+					// wait for chain to change
+					const onceListener = this.network.networkingReady.subscribe((value) => {
+						if (value) {
+							event.sender.send('changeChainResponse', result);
+							onceListener.unsubscribe();
+						}
+					});
+				} else {
+					event.sender.send('changeChainResponse', true);
+				}
+				break;
+			}
+			case 'authorizations': {
+				this.toaster.onInfo('wallet connection', 'external application requested public account info');
+				this.transitconnect = true;
+				this.isSimpleosConnect = true;
 
-    onSignModalClose(ev) {
-        if (this.transitEventHandler && ev === false && this.eventFired === false) {
-            this.eventFired = true;
-            this.transitEventHandler.sender.send('signResponse', {status: 'CANCELLED'});
-        }
-        if (this.replyEvent && ev === false && this.eventFired === false) {
-            this.eventFired = true;
-            this.replyEvent.sender.send('signResponse', {status: 'CANCELLED'});
-        }
-    }
+				if (this.aService.accounts.length > 0) {
+					this.accountChange = this.selectedAccount.subscribe((data) => {
+						if (data) {
+							event.sender.send('authorizationsResponse', data);
+							this.accountChange.unsubscribe();
+							this.selectedAccount.next(null);
+							this.transitconnect = false;
+						}
+					});
+				} else {
+					console.log('No account found!');
+					event.sender.send('authorizationsResponse', {});
+				}
 
-    toggleDayNight() {
-        this.dnSet = !this.dnSet;
-        if (this.dnSet) {
-            localStorage.setItem('use_light_theme', 'true');
-            this.theme.lightMode = true;
-            this.theme.lightTheme();
-        } else {
-            localStorage.setItem('use_light_theme', 'false');
-            this.theme.lightMode = false;
-            if (this.compilerVersion === 'LIBERLAND') {
-                this.theme.liberlandTheme();
-            } else {
-                this.theme.defaultTheme();
-            }
-        }
-        this.cdr.detectChanges();
-    }
+				break;
+			}
+			case 'sign': {
+				if (this.crypto.getLockStatus()) {
+					event.sender.send('signResponse', {
+						status: 'CANCELLED'
+					});
+				} else {
+					this.simpleosConnectSign(event, payload).catch(console.log);
+				}
+				break;
+			}
+		}
+	}
 
-    public minimizeWindow() {
-        console.log('Minimize...');
-        if (this._electronService.isElectronApp) {
-            this._electronService.remote.getCurrentWindow().minimize();
-        }
-    }
+	private changeChain(chainId): boolean {
+		if (this.network.defaultChains.find((chain) => chain.id === chainId)) {
+			if (this.network.activeChain.id !== chainId) {
+				this.network.changeChain(chainId).catch(console.log);
+			}
+			return true;
+		} else {
+			return false;
+		}
+	}
 
-    public closeWindow() {
-        console.log('Close...');
-        if (this._electronService.isElectronApp) {
-            this._electronService.remote.getCurrentWindow().close();
-        }
-    }
+	onLoginModalClose(ev) {
+		if (this.transitEventHandler && ev === false && this.eventFired === false) {
+			this.eventFired = true;
+			this.transitEventHandler.sender.send('loginResponse', {status: 'CANCELLED'});
+		}
+	}
 
-    public maximizeWindow() {
-        console.log('Maximize...');
-        if (this._electronService.isElectronApp) {
-            if (this._electronService.remote.getCurrentWindow().isMaximized()) {
-                this._electronService.remote.getCurrentWindow().restore();
-            } else {
-                this._electronService.remote.getCurrentWindow().maximize();
-            }
-        }
-    }
+	onSignModalClose(ev) {
+		if (this.transitEventHandler && ev === false && this.eventFired === false) {
+			this.eventFired = true;
+			this.transitEventHandler.sender.send('signResponse', {status: 'CANCELLED'});
+		}
+		if (this.replyEvent && ev === false && this.eventFired === false) {
+			this.eventFired = true;
+			this.replyEvent.sender.send('signResponse', {status: 'CANCELLED'});
+		}
+	}
 
-    get maximized(): boolean {
-        return this._electronService.remote.getCurrentWindow().isMaximized();
-    }
+	toggleDayNight() {
+		this.dnSet = !this.dnSet;
+		if (this.dnSet) {
+			localStorage.setItem('use_light_theme', 'true');
+			this.theme.lightMode = true;
+			this.theme.lightTheme();
+		} else {
+			localStorage.setItem('use_light_theme', 'false');
+			this.theme.lightMode = false;
+			if (this.compilerVersion === 'LIBERLAND') {
+				this.theme.liberlandTheme();
+			} else {
+				this.theme.defaultTheme();
+			}
+		}
+		this.cdr.detectChanges();
+	}
 
-    mode: string = 'local';
+	public minimizeWindow() {
+		console.log('Minimize...');
+		if (this._electronService.isElectronApp) {
+			this._electronService.remote.getCurrentWindow().minimize();
+		}
+	}
 
-    performUpdate() {
-        if (this.compilerVersion === 'LIBERLAND') {
-            if (this.newVersion['homepage']) {
-                window['shell'].openExternal(this.newVersion['homepage']).catch(console.log);
-            } else {
-                this.openGithub();
-            }
-        } else {
-            window['shell'].openExternal('https://eosrio.io/simpleos/').catch(console.log);
-        }
-    }
+	public closeWindow() {
+		console.log('Close...');
+		if (this._electronService.isElectronApp) {
+			this._electronService.remote.getCurrentWindow().close();
+		}
+	}
 
-    openGithub() {
-        window['shell'].openExternal(this.newVersion['link']).catch(console.log);
-    }
+	public maximizeWindow() {
+		console.log('Maximize...');
+		if (this._electronService.isElectronApp) {
+			if (this._electronService.remote.getCurrentWindow().isMaximized()) {
+				this._electronService.remote.getCurrentWindow().restore();
+			} else {
+				this._electronService.remote.getCurrentWindow().maximize();
+			}
+		}
+	}
 
-    selectAccount(account_data, idx) {
+	performUpdate() {
+		if (this.compilerVersion === 'LIBERLAND') {
+			if (this.newVersion['homepage']) {
+				window['shell'].openExternal(this.newVersion['homepage']).catch(console.log);
+			} else {
+				this.openGithub();
+			}
+		} else {
+			window['shell'].openExternal('https://eosrio.io/simpleos/').catch(console.log);
+		}
+	}
 
-        const [auth, key] = this.trxFactory.getAuth(account_data);
-        let responseData;
-        if (key !== '') {
-            if (this.isSimpleosConnect) {
-                responseData = {
-                    actor: account_data.name,
-                    permission: auth.permission,
-                    key: key,
-                };
-            } else {
-                responseData = {
-                    accountName: account_data.name,
-                    permission: auth.permission,
-                    publicKey: key,
-                };
-                this.transitconnect = false;
-            }
-            this.selectedAccount.next(responseData);
-            this.aService.select(idx);
-        }
-    }
+	openGithub() {
+		window['shell'].openExternal(this.newVersion['link']).catch(console.log);
+	}
 
-    signResponse(data) {
-        this.replyEvent.sender.send('signResponse', data);
-    }
+	selectAccount(account_data, idx) {
 
-    async signExternalAction() {
-        this.wrongpass = '';
-        this.txerror = '';
-        this.busy = true;
+		const [auth, key] = this.trxFactory.getAuth(account_data);
+		let responseData;
+		if (key !== '') {
+			if (this.isSimpleosConnect) {
+				responseData = {
+					actor: account_data.name,
+					permission: auth.permission,
+					key: key,
+				};
+			} else {
+				responseData = {
+					accountName: account_data.name,
+					permission: auth.permission,
+					publicKey: key,
+				};
+				this.transitconnect = false;
+			}
+			this.selectedAccount.next(responseData);
+			this.aService.select(idx);
+		}
+	}
 
-        if (this.mode === 'local') {
-            console.log('signing in local mode');
+	signResponse(data) {
+		this.replyEvent.sender.send('signResponse', data);
+	}
 
-            const account = this.aService.accounts.find(a => a.name === this.external_signer);
-            const idx = this.aService.accounts.indexOf(account);
-            this.aService.selected.next(account);
-            const [, publicKey] = this.trxFactory.getAuth();
+	async signExternalAction() {
+		this.wrongpass = '';
+		this.txerror = '';
+		this.busy = true;
 
-            try {
-                await this.crypto.authenticate(this.confirmForm.get('pass').value, publicKey);
-            } catch (e) {
-                this.wrongpass = 'wrong password';
-                this.busy = false;
-                return;
-            }
+		if (this.mode === 'local') {
+			console.log('signing in local mode');
 
-            try {
+			const account = this.aService.accounts.find(a => a.name === this.external_signer);
+			const idx = this.aService.accounts.indexOf(account);
+			this.aService.selected.next(account);
+			const [, publicKey] = this.trxFactory.getAuth();
 
-                console.log(this.fullTrxData, !this.transitMode);
-                const result = await this.eosjs.signTrx(this.fullTrxData, !this.transitMode);
-                console.log(result);
+			try {
+				await this.crypto.authenticate(this.confirmForm.get('pass').value, publicKey);
+			} catch (e) {
+				this.wrongpass = 'wrong password';
+				this.busy = false;
+				return;
+			}
 
-                if (result) {
-                    if (this.transitMode) {
-                        this.signResponse({
-                            sigs: result.packedTransaction.signatures,
-                        });
-                    } else {
-                        this.signResponse({
-                            status: 'OK',
-                            content: result.result,
-                            packed: result.packedTransaction
-                        });
-                    }
-                    this.wrongpass = '';
-                    this.txerror = '';
-                    this.busy = false;
-                    this.externalActionModal = false;
-                    this.aService.select(idx);
-                    this.aService.reloadActions(account);
-                    await this.aService.refreshFromChain(false);
-                    this.cdr.detectChanges();
-                }
-            } catch (e) {
-                this.txerror = e;
-                console.log(e);
-                this.busy = false;
-            }
+			try {
 
-        } else if (this.mode === 'ledger') {
-            try {
-                const result = await this.ledger.sign(
-                    this.fullTrxData,
-                    this.crypto.requiredLedgerSlot,
-                    this.network.selectedEndpoint.getValue().url
-                );
-                if (result) {
-                    this.signResponse({
-                        status: 'OK',
-                        content: result.result,
-                        packed: result.packedTransaction
-                    });
-                    this.wrongpass = '';
-                    this.txerror = '';
-                    this.busy = false;
-                    this.externalActionModal = false;
-                    this.cdr.detectChanges();
-                }
-            } catch (e) {
-                this.signResponse({
-                    status: 'ERROR',
-                    reason: e
-                });
-                this.txerror = e;
-                console.log(e);
-                this.busy = false;
-            }
-        }
-    }
+				console.log(this.fullTrxData, !this.transitMode);
+				const result = await this.eosjs.signTrx(this.fullTrxData, !this.transitMode);
+				console.log(result);
 
-    processSigners() {
-        let signer = '';
-        for (const action of this.fullTrxData.actions) {
-            if (action.account === 'eosio' && action.name === 'updateauth') {
-                this.updateauthWarning = true;
-            }
-            if (signer === '') {
-                signer = action.authorization[0].actor;
-            } else {
-                if (signer !== action.authorization[0].actor) {
-                    console.log('Multiple signers!!!');
-                }
-            }
-        }
-        this.external_signer = signer;
-    }
+				if (result) {
+					if (this.transitMode) {
+						this.signResponse({sigs: result.packedTransaction.signatures});
+					} else {
+						this.signResponse({
+							status: 'OK',
+							content: result.result,
+							packed: result.packedTransaction
+						});
+					}
+					this.wrongpass = '';
+					this.txerror = '';
+					this.busy = false;
+					this.externalActionModal = false;
+					this.aService.select(idx);
+					this.aService.reloadActions(account);
+					await this.aService.refreshFromChain(false);
+					this.cdr.detectChanges();
+				}
+			} catch (e) {
+				this.txerror = e;
+				console.log(e);
+				this.busy = false;
+			}
 
-    async onTransitSign(event, payload) {
-        if (this.crypto.getLockStatus()) {
-            return;
-        }
-        this.loadingTRX = true;
-        const hexData = payload.content.hex_data;
-        try {
-            const data = await this.eosjs.localSigProvider.processTrx(hexData);
-            this.fullTrxData = data;
-            this.updateauthWarning = false;
-            this.confirmForm.reset();
-            this.processSigners();
-            this.action_json = data.actions;
-            this.zone.run(() => {
-                this.loadingTRX = false;
-                this.externalActionModal = true;
-                this.eventFired = false;
-            });
-            this.replyEvent = event;
-        } catch (e) {
-            console.log(e);
-            this.loadingTRX = false;
-        }
-    }
+		} else if (this.mode === 'ledger') {
+			try {
+				const result = await this.ledger.sign(
+					this.fullTrxData,
+					this.crypto.requiredLedgerSlot,
+					this.network.selectedEndpoint.getValue().url
+				);
+				if (result) {
+					this.signResponse({
+						status: 'OK',
+						content: result.result,
+						packed: result.packedTransaction
+					});
+					this.wrongpass = '';
+					this.txerror = '';
+					this.busy = false;
+					this.externalActionModal = false;
+					this.cdr.detectChanges();
+				}
+			} catch (e) {
+				this.signResponse({
+					status: 'ERROR',
+					reason: e
+				});
+				this.txerror = e;
+				console.log(e);
+				this.busy = false;
+			}
+		}
+	}
 
-    async simpleosConnectSign(event, payload) {
-        this.fullTrxData = payload.content;
-        this.loadingTRX = true;
-        try {
-            this.updateauthWarning = false;
-            this.confirmForm.reset();
+	processSigners() {
+		let signer = '';
+		for (const action of this.fullTrxData.actions) {
+			if (action.account === 'eosio' && action.name === 'updateauth') {
+				this.updateauthWarning = true;
+			}
+			if (signer === '') {
+				signer = action.authorization[0].actor;
+			} else {
+				if (signer !== action.authorization[0].actor) {
+					console.log('Multiple signers!!!');
+				}
+			}
+		}
+		this.external_signer = signer;
+	}
 
-            this.processSigners();
-            this.extendActionJson();
-            this.checkSignerMode();
+	async onTransitSign(event, payload) {
+		if (this.crypto.getLockStatus()) {
+			return;
+		}
+		this.loadingTRX = true;
+		const hexData = payload.content.hex_data;
+		try {
+			const data = await this.eosjs.localSigProvider.processTrx(hexData);
+			this.fullTrxData = data;
+			this.updateauthWarning = false;
+			this.confirmForm.reset();
+			this.processSigners();
+			this.action_json = data.actions;
+			this.zone.run(() => {
+				this.loadingTRX = false;
+				this.externalActionModal = true;
+				this.eventFired = false;
+			});
+			this.replyEvent = event;
+		} catch (e) {
+			console.log(e);
+			this.loadingTRX = false;
+		}
+	}
 
-            // assign reply event
-            this.replyEvent = event;
+	async simpleosConnectSign(event, payload) {
+		this.fullTrxData = payload.content;
+		this.loadingTRX = true;
+		try {
+			this.updateauthWarning = false;
+			this.confirmForm.reset();
 
-            this.zone.run(() => {
-                this.wrongpass = '';
-                this.loadingTRX = false;
-                this.externalActionModal = true;
-                this.eventFired = false;
-            });
-        } catch (e) {
-            console.log(e);
-            this.loadingTRX = false;
-        }
-    }
+			this.processSigners();
+			this.extendActionJson();
+			this.checkSignerMode();
 
-    private checkSignerMode() {
-        const account = this.aService.accounts.find(a => a.name === this.external_signer);
-        const [, publicKey] = this.trxFactory.getAuth(account);
-        this.mode = this.crypto.getPrivateKeyMode(publicKey);
-    }
+			// assign reply event
+			this.replyEvent = event;
 
-    get requiredLedgerInfo() {
-        return {
-            device: this.crypto.requiredLedgerDevice,
-            slot: this.crypto.requiredLedgerSlot
-        };
-    }
+			this.zone.run(() => {
+				this.wrongpass = '';
+				this.loadingTRX = false;
+				this.externalActionModal = true;
+				this.eventFired = false;
+			});
+		} catch (e) {
+			console.log(e);
+			this.loadingTRX = false;
+		}
+	}
 
-    private extendActionJson() {
-        // deep clone transaction data for display
-        this.action_json = JSON.parse(JSON.stringify(this.fullTrxData.actions));
-        for (const act of this.action_json) {
-            if (act.data) {
-                for (const k in act.data) {
-                    if (act.data.hasOwnProperty(k)) {
-                        try {
-                            act.data[k] = JSON.parse(act.data[k]);
-                        } catch (e) {
+	private checkSignerMode() {
+		const account = this.aService.accounts.find(a => a.name === this.external_signer);
+		const [, publicKey] = this.trxFactory.getAuth(account);
+		this.mode = this.crypto.getPrivateKeyMode(publicKey);
+	}
 
-                        }
-                    }
-                }
-            }
-        }
-    }
+	private extendActionJson() {
+		// deep clone transaction data for display
+		this.action_json = JSON.parse(JSON.stringify(this.fullTrxData.actions));
+		for (const act of this.action_json) {
+			if (act.data) {
+				for (const k in act.data) {
+					if (act.data.hasOwnProperty(k)) {
+						try {
+							act.data[k] = JSON.parse(act.data[k]);
+						} catch (e) {
+
+						}
+					}
+				}
+			}
+		}
+	}
 }
